+++ lib/streamlit/elements/number_input.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 import numbers
-from streamlit.scriptrunner import ScriptRunContext, get_script_run_ctx
+from streamlit.runtime.scriptrunner import ScriptRunContext, get_script_run_ctx
 from streamlit.type_util import Key, to_key
 from textwrap import dedent
 from typing import Optional, Union, cast
