+++ lib/streamlit/elements/time_widgets.py
@@ -20,7 +20,7 @@ from typing import Any, cast, List, Optional, Sequence, Tuple, TYPE_CHECKING, Un
 from dateutil import relativedelta
 from typing_extensions import TypeAlias
 
-from streamlit.scriptrunner import ScriptRunContext, get_script_run_ctx
+from streamlit.runtime.scriptrunner import ScriptRunContext, get_script_run_ctx
 from streamlit.type_util import Key, to_key
 from streamlit.errors import StreamlitAPIException
 from streamlit.proto.DateInput_pb2 import DateInput as DateInputProto
