+++ lib/streamlit/in_memory_file_manager.py
@@ -40,7 +40,7 @@ FILE_TYPE_DOWNLOADABLE = "downloadable_file"
 
 def _get_session_id() -> str:
     """Semantic wrapper to retrieve current AppSession ID."""
-    from streamlit.scriptrunner import get_script_run_ctx
+    from streamlit.runtime.scriptrunner import get_script_run_ctx
 
     ctx = get_script_run_ctx()
     if ctx is None:
