+++ lib/streamlit/user_info.py
@@ -1,8 +1,8 @@
 from typing import Iterator, Mapping, NoReturn, Optional
 
 from streamlit.errors import StreamlitAPIException
-from streamlit.scriptrunner import get_script_run_ctx as _get_script_run_ctx
-from streamlit.scriptrunner.script_run_context import UserInfo
+from streamlit.runtime.scriptrunner import get_script_run_ctx as _get_script_run_ctx
+from streamlit.runtime.scriptrunner.script_run_context import UserInfo
 
 
 def _get_user_info() -> UserInfo:
