+++ lib/tests/streamlit/file_uploader_test.py
@@ -19,7 +19,7 @@ from unittest.mock import patch
 
 import streamlit as st
 from streamlit import config
-from streamlit.scriptrunner import get_script_run_ctx
+from streamlit.runtime.scriptrunner import get_script_run_ctx
 from streamlit.uploaded_file_manager import UploadedFileRec, UploadedFile
 from tests import testutil
