+++ lib/tests/streamlit/runtime/scriptrunner/test_data/infinite_loop.py
