+++ lib/tests/streamlit/script_run_context_test.py
@@ -16,7 +16,7 @@ import unittest
 
 from streamlit.errors import StreamlitAPIException
 from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
-from streamlit.scriptrunner import ScriptRunContext
+from streamlit.runtime.scriptrunner import ScriptRunContext
 from streamlit.state import SafeSessionState, SessionState
 from streamlit.uploaded_file_manager import UploadedFileManager
