+++ lib/tests/streamlit/state/session_state_test.py
@@ -26,7 +26,7 @@ import streamlit as st
 from streamlit.errors import StreamlitAPIException
 from streamlit.proto.WidgetStates_pb2 import WidgetState as WidgetStateProto
 from streamlit.proto.WidgetStates_pb2 import WidgetStates as WidgetStatesProto
-from streamlit.scriptrunner import get_script_run_ctx
+from streamlit.runtime.scriptrunner import get_script_run_ctx
 from streamlit.state.session_state_proxy import get_session_state
 from streamlit.state.session_state import (
     GENERATED_WIDGET_KEY_PREFIX,
@@ -540,7 +540,9 @@ class SessionStateMethodTests(unittest.TestCase):
         mock_ctx = MagicMock()
         mock_ctx.widget_ids_this_run = {"widget_id"}
 
-        with patch("streamlit.scriptrunner.get_script_run_ctx", return_value=mock_ctx):
+        with patch(
+            "streamlit.runtime.scriptrunner.get_script_run_ctx", return_value=mock_ctx
+        ):
             with pytest.raises(StreamlitAPIException) as e:
                 self.session_state._key_id_mapping = {"widget_id": "widget_id"}
                 self.session_state["widget_id"] = "blah"
@@ -550,7 +552,9 @@ class SessionStateMethodTests(unittest.TestCase):
         mock_ctx = MagicMock()
         mock_ctx.form_ids_this_run = {"form_id"}
 
-        with patch("streamlit.scriptrunner.get_script_run_ctx", return_value=mock_ctx):
+        with patch(
+            "streamlit.runtime.scriptrunner.get_script_run_ctx", return_value=mock_ctx
+        ):
             with pytest.raises(StreamlitAPIException) as e:
                 self.session_state["form_id"] = "blah"
             assert "`st.session_state.form_id` cannot be modified" in str(e.value)
